pyproject.toml
[project]
name = "qtype"
version = "0.1.0"
description = "DSL for Generative AI Prototyping"
authors = [{ name="Lou Kratz", email="lou.kratz+qtype@bazaarvoice.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jsonschema>=4.24.0",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.2",
    "python-dotenv>=1.0.0",
    "openai>=1.93.0",
    "fsspec>=2025.5.1",
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "pip-system-certs>=5.2",
    "openapi3-parser>=1.1.21",
    "pydantic-yaml>=1.6.0",
    "google-cloud-aiplatform>=1.120.0",
]
license = "APACHE-2.0"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/bazaarvoice/qtype"

[project.optional-dependencies]
interpreter = [
    "aiostream>=0.7.1",
    "arize-phoenix-otel>=0.12.1",
    "boto3>=1.34.0",
    "datasets>=4.4.1",
    "diskcache>=5.6.3",
    "docling>=2.55.1",
    "docx2txt>=0.9",
    "fastapi>=0.116.1",
    "jsonpath-ng>=1.7.0",
    "langfuse>=3.9.0",
    "llama-index-embeddings-bedrock>=0.5.2",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-llms-bedrock-converse>=0.10.5",
    "llama-index-llms-bedrock>=0.3.8",
    "llama-index-llms-vertex>=0.6.1",
    "llama-index-postprocessor-bedrock-rerank>=0.5.1",
    "llama-index-readers-huggingface-fs>=0.4.1",
    "llama-index-vector-stores-qdrant>=0.8.6",
    "llama-index>=0.12.45",
    "openinference-instrumentation-llama-index>=4.3.4",
    "opensearch-py>=2.7.0",
    "opentelemetry-exporter-otlp>=1.35.0",
    "opentelemetry-sdk>=1.35.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pyarrow>=21.0.0",
    "pyathena[sqlalchemy]>=3.18.0",
    "python-magic>=0.4.27",
    "s3fs>=2025.7.0",
    "sqlalchemy>=2.0.42",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "arize-phoenix>=11.2.2",
    "boto3>=1.34.0",
    "coverage>=7.0.0",
    "ipython>=8.37.0",
    "isort>=5.13.0",
    "mermaid-py>=0.8.0",
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "mkdocs-material>=9.6.15",
    "mkdocs>=1.6.1",
    "mkdocstrings-python>=1.16.12",
    "mkdocstrings>=0.30.0",
    "networkx>=3.4.2",
    "pandas-stubs>=2.3.2.250827",
    "pkginfo>=1.12.1.2",
    "pre-commit>=3.6.0",
    "pymdown-extensions>=10.16",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.0.0",
    "pytest>=8.4.1",
    "ruff>=0.1.0",
    "types-cachetools>=6.2.0.20250827",
    "types-networkx>=3.5.0.20250901",
    "types-PyYAML>=6.0.2",
    "types-requests>=2.32.4.20250809",
    "ty>=0.0.8",
]
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
]

[tool.uv]
# Install dev dependencies by default when running uv sync
default-groups = ["dev"]
# Install the package itself in editable mode
package = true
# Use highest resolution strategy for dependencies  
resolution = "highest"
# Compile bytecode for faster subsequent runs
compile-bytecode = true

[project.scripts]
qtype = "qtype.cli:main"

[tool.setuptools.packages.find]
include = ["qtype*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
"qtype.interpreter" = ["ui/**/*"]

# Tool configurations
[tool.ruff]
line-length = 79
target-version = "py310"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]  # Let ruff handle line length

[tool.isort]
line_length = 79
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# Note: ty will show references to .venv in error contexts when reporting
# type issues with imported external libraries. This is expected behavior.
[tool.ty.environment]
root = ["."]

[tool.ty.src]
include = ["qtype/**"]

[tool.coverage.run]
source = ["qtype"]
omit = [
    "qtype/commands/*",
    "qtype/interpreter/*", # TODO: expand coverage here.
    "qtype/*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "network: marks tests as requiring network access",
    "asyncio: marks tests as async tests using asyncio",
]
# Configure pytest-asyncio to automatically detect async tests
asyncio_mode = "auto"
